pyproject.toml
[project]
name = "groq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.0.0",
    "groq",
    "python-dotenv",
    "mcp[cli]",
]
